requirements.txt
asgiref==3.8.1
Django==5.1.2
djangorestframework==3.15.2
inflection==0.5.1
pillow==10.4.0
psycopg2-binary==2.9.9
PyYAML==6.0.2
sqlparse==0.5.1
uritemplate==4.1.1
